split up io.vector in submodules; fix broken imports; improve typing
