+++ mapchete/commands/convert.py
@@ -20,7 +20,7 @@ from mapchete.errors import JobCancelledError
 from mapchete.executor import Executor
 from mapchete.formats import available_output_formats
 from mapchete.io import MPath, fiona_open, get_best_zoom_level
-from mapchete.io.vector import reproject_geometry
+from mapchete.geometry import reproject_geometry
 from mapchete.tile import BufferedTilePyramid
 from mapchete.types import MPathLike, ResamplingLike, to_resampling
