+++ mapchete/geometry/shape.py
@@ -1,9 +1,10 @@
+from typing import Any
 from shapely.geometry import shape
 
-from mapchete.geometry.types import GeoInterface, Geometry, GeometryLike
+from mapchete.geometry.types import GeoInterface, Geometry
 
 
-def to_shape(geometry: GeometryLike) -> Geometry:
+def to_shape(geometry: Any) -> Geometry:
     """
     Convert geometry to shapely geometry if necessary.
 
@@ -18,6 +19,8 @@ def to_shape(geometry: GeometryLike) -> Geometry:
     try:
         if isinstance(geometry, Geometry):
             return geometry
+        elif isinstance(geometry, dict) and geometry.get("geometry"):
+            return shape(geometry["geometry"])
         elif (
             isinstance(geometry, GeoInterface)
             and isinstance(geometry.__geo_interface__, dict)
@@ -25,6 +28,6 @@ def to_shape(geometry: GeometryLike) -> Geometry:
         ):
             return shape(geometry.__geo_interface__["geometry"])
         else:
-            return shape(geometry)
+            return shape(geometry)  # type: ignore
     except Exception:  # pragma: no cover
         raise TypeError(f"invalid geometry type: {type(geometry)}")
