+++ mapchete/io/vector/__init__.py
@@ -1,0 +1,20 @@
+from mapchete.io.vector.open import fiona_open
+from mapchete.io.vector.read import (
+    IndexedFeatures,
+    convert_vector,
+    fiona_read,
+    read_vector,
+    read_vector_window,
+)
+from mapchete.io.vector.write import fiona_write, write_vector_window
+
+__all__ = [
+    "fiona_read",
+    "fiona_write",
+    "fiona_open",
+    "read_vector_window",
+    "write_vector_window",
+    "IndexedFeatures",
+    "convert_vector",
+    "read_vector",
+]
